boards/arm/stm32f7: move CAN init logic to a common directory
move CAN init logic to a common directory
