+++ boards/arm/stm32f7/nucleo-f722ze/configs/can/defconfig
@@ -8,6 +8,7 @@
 # CONFIG_ARCH_FPU is not set
 # CONFIG_STM32F7_USE_LEGACY_PINMAP is not set
 CONFIG_ARCH="arm"
+CONFIG_ARCH_BOARD_COMMON=y
 CONFIG_ARCH_BOARD="nucleo-f722ze"
 CONFIG_ARCH_BOARD_NUCLEO_F722ZE=y
 CONFIG_ARCH_BUTTONS=y
