+++ boards/arm/stm32f7/nucleo-f722ze/configs/cansock/defconfig
@@ -10,6 +10,7 @@
 # CONFIG_NET_IPv4 is not set
 # CONFIG_STM32F7_USE_LEGACY_PINMAP is not set
 CONFIG_ARCH="arm"
+CONFIG_ARCH_BOARD_COMMON=y
 CONFIG_ARCH_BOARD="nucleo-f722ze"
 CONFIG_ARCH_BOARD_NUCLEO_F722ZE=y
 CONFIG_ARCH_BUTTONS=y
