+++ boards/arm/stm32f7/nucleo-f746zg/configs/pysim/defconfig
@@ -12,6 +12,7 @@ CONFIG_ADC=y
 CONFIG_ADC_FIFOSIZE=16
 CONFIG_ANALOG=y
 CONFIG_ARCH="arm"
+CONFIG_ARCH_BOARD_COMMON=y
 CONFIG_ARCH_BOARD="nucleo-f746zg"
 CONFIG_ARCH_BOARD_NUCLEO_F746ZG=y
 CONFIG_ARCH_BUTTONS=y
