+++ boards/arm/stm32f7/nucleo-f746zg/src/CMakeLists.txt
@@ -74,15 +74,6 @@ if(CONFIG_SENSORS_QENCODER)
   list(APPEND SRCS stm32_qencoder.c)
 endif()
 
-if(CONFIG_STM32F7_CAN)
-  if(CONFIG_STM32F7_CAN_CHARDRIVER)
-    list(APPEND SRCS stm32_can.c)
-  endif()
-  if(CONFIG_STM32F7_CAN_SOCKET)
-    list(APPEND SRCS stm32_cansock.c)
-  endif()
-endif()
-
 if(CONFIG_USBDEV_COMPOSITE)
   list(APPEND SRCS stm32_composite.c)
 endif()
