+++ boards/arm/stm32f7/nucleo-f746zg/src/Make.defs
@@ -76,15 +76,6 @@ ifeq ($(CONFIG_SENSORS_QENCODER),y)
 CSRCS += stm32_qencoder.c
 endif
 
-ifeq ($(CONFIG_STM32F7_CAN),y)
-ifeq ($(CONFIG_STM32F7_CAN_CHARDRIVER),y)
-CSRCS += stm32_can.c
-endif
-ifeq ($(CONFIG_STM32F7_CAN_SOCKET),y)
-CSRCS += stm32_cansock.c
-endif
-endif
-
 ifeq ($(CONFIG_USBDEV_COMPOSITE),y)
 CSRCS += stm32_composite.c
 endif
