+++ boards/arm/stm32f7/nucleo-f746zg/src/nucleo-f746zg.h
@@ -326,22 +326,6 @@ int stm32_qencoder_initialize(const char *devpath, int timer);
 #endif
 
 /****************************************************************************
- * Name: stm32_can_setup
- ****************************************************************************/
-
-#ifdef CONFIG_STM32F7_CAN_CHARDRIVER
-int stm32_can_setup(void);
-#endif
-
-/****************************************************************************
- * Name: stm32_cansock_setup
- ****************************************************************************/
-
-#ifdef CONFIG_STM32F7_CAN_SOCKET
-int stm32_cansock_setup(void);
-#endif
-
-/****************************************************************************
  * Name: stm32f7_gpio_initialize
  ****************************************************************************/
